2025-07-29_LocalStack_Community_Meetup/generated/requirements.txt
flask
snowflake-connector-python 